WIP: finish add fund endpoint
The first approach was to add a use of case called CreateFund that use FundsCreator and FundFinder but there is less code by using the repository directly in FundsCreator along with the FundFinder.

If there is some place where fund must be created, FundsCreator could be split in another class and the current FundsCreator could have the CreateFund name and make use of FundsCreate along FundFinder instead of using the repository directly.